requirements.txt
playwright
beautifulsoup4
python-docx
requests
lxml
selenium
webdriver-manager
flask
